requirements.txt
click==8.2.1
rich==13.7.1
requests==2.32.3
typing-extensions==4.12.2
rapidfuzz==3.9.7
pydantic==2.8.2
pandas==2.2.2
numpy==1.26.4
fhir.resources==7.1.0
transformers==4.41.2
tokenizers==0.19.1
accelerate==0.33.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
scikit-learn==1.5.1
python-dotenv==1.0.1
uvicorn==0.30.6
fastapi==0.112.0
# OCR stack
torch==2.4.1
pymupdf==1.24.9
Pillow==10.4.0
opencv-python-headless==4.10.0.84
python-bidi==0.4.2
# Surya OCR (pulled via pip; version may resolve to latest compatible)
surya-ocr==0.6.2
# Evaluation / RAG add‑ons
ragas==0.1.13
datasets==2.20.0
pdfminer.six==20240706
# Optional Google GenAI connectors
google-generativeai==0.8.3
langchain-google-genai==2.0.5
